+++ src/structures/Weights.hpp
@@ -91,7 +91,7 @@ class Weights {
     * only the active regions of the patches are considered when taking the minimum. */
    float calcMaxWeight(int arbor);
 
-   int calcDataIndexFromPatchIndex(int patchIndex) const;
+   long calcDataIndexFromPatchIndex(long patchIndex) const;
 
 #ifdef PV_USE_CUDA
    /**
